+++ extensions/chocolatey-vscode.extension/CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
-## 1.0
+## 1.1.0
+
+- `Install-VsCodeExtension` supports now also Insider builds and user level installer
+- `Uninstall-VsCodeExtension` supports now also Insider builds and user level installer
+
+## 1.0.0
 
 - Added `Install-VsCodeExtension`
 - Added `Uninstall-VsCodeExtension`